test: count the clock over the exposure it was read across
The regenerated documents inherited their spacecraft clock strings as
typed digits, so each triple disagreed with the epochs beside it: a
Cassini exposure declaring 0.46 seconds spanned twenty-five ticks of the
256-per-second clock, about a tenth of that, and Voyager was out by
twenty-three. A real triple comes from the epochs through the mission
clock kernel, so one that contradicts them is the same kind of invention
the rest of this corpus just stopped making.

Both readings are now counted from the epochs actually written, in the
spelling each mission's kernel defines: Cassini's zero-padded seconds and
1/256 fraction, Voyager's FDS count, minor frame and line, each carrying
at its own modulus. Every span now agrees with its exposure to within a
quarter of one tick, and a test re-derives that from the stored strings
rather than asking the builders.

The generator becomes a package to stay under the module cap, one module
per host, and the tests drop docstring parameters for fixtures they are
handed.
