+++ tests/spindoctor/cli/stats/data/results_tree/COISS_2001/data/1294561143_1295221348/N1294562000_1_CALIB_metadata.json
@@ -136,8 +136,8 @@
       "midtime_et": 170000800.0,
       "exposure_s": 0.46,
       "sclk_start": "1/1294562000.055",
-      "sclk_midtime": "1/1294562000.062",
-      "sclk_stop": "1/1294562000.070"
+      "sclk_midtime": "1/1294562000.114",
+      "sclk_stop": "1/1294562000.173"
     }
   },
   "timing": {
