+++ tests/spindoctor/cli/stats/data/results_tree/COISS_2001/data/1294561143_1295221348/N1294564000_1_CALIB_metadata.json
@@ -424,8 +424,8 @@
       "midtime_et": 170002800.0,
       "exposure_s": 0.46,
       "sclk_start": "1/1294564000.011",
-      "sclk_midtime": "1/1294564000.030",
-      "sclk_stop": "1/1294564000.049"
+      "sclk_midtime": "1/1294564000.070",
+      "sclk_stop": "1/1294564000.129"
     }
   },
   "timing": {
