+++ tests/spindoctor/cli/stats/data/results_tree/COISS_2001/data/1294561143_1295221348/W1294561202_1_CALIB_metadata.json
@@ -201,8 +201,8 @@
       "midtime_et": 170000000.0,
       "exposure_s": 0.46,
       "sclk_start": "1/1294561202.077",
-      "sclk_midtime": "1/1294561202.089",
-      "sclk_stop": "1/1294561202.102"
+      "sclk_midtime": "1/1294561202.136",
+      "sclk_stop": "1/1294561202.195"
     }
   },
   "timing": {
